Create MAAT met driver files

met2model wrapper for MAAT

Usage

met2model.MAAT(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  leap_year = TRUE,
  ...
)

Arguments

in.path

location on disk where inputs (CF met drivers) are stored

in.prefix

prefix of input and output files

outfolder

location on disk where MAAT met outputs will be stored

start_date

the start date of the data to be downloaded (will only use the year part of the date)

end_date

the end date of the data to be downloaded (will only use the year part of the date)

overwrite

should existing files be overwritten

verbose

should the function be very verbose

leap_year

Enforce Leap-years? If set to TRUE, will require leap years to have 366 days. If set to false, will require all years to have 365 days. Default = TRUE.

...

additional arguments, currently ignored

Author

Shawn P. Serbin