Writes a MAAT config file.

Requires a pft xml object, a list of trait values for a single model run, and the name of the file to create

Usage

write.config.MAAT(defaults = NULL, trait.values, settings, run.id)

Arguments

defaults

list of defaults to process

trait.values

vector of samples for a given trait

settings

list of settings from pecan settings file

run.id

id of run

Value

configuration file for MAAT for given run

Author

Shawn Serbin, Anthony Walker, Rob Kooper, Chris Black